+++ blas/cuda/NativeOps.cu
@@ -4299,7 +4299,7 @@ void NativeOps::pullRowsFloat(Nd4jPointer *extraPointers, Nd4jPointer x, Nd4jPoi
 	int *zShape = reinterpret_cast<int *>(zShapeInfo);
 	int *xShape = reinterpret_cast<int *>(xShapeInfo);
 
-	long *index = reinterpret_cast<long *>(indexes);
+	int *index = reinterpret_cast<int *>(indexes);
 	int *tadOnlyShapeInfo = reinterpret_cast<int *>(tadShapeInfo);
 	int *tadOffset = reinterpret_cast<int *>(tadOffsets);
 
@@ -4315,7 +4315,7 @@ void NativeOps::pullRowsDouble(Nd4jPointer *extraPointers, Nd4jPointer x, Nd4jPo
 	int *zShape = reinterpret_cast<int *>(zShapeInfo);
 	int *xShape = reinterpret_cast<int *>(xShapeInfo);
 
-	long *index = reinterpret_cast<long *>(indexes);
+	int *index = reinterpret_cast<int *>(indexes);
 	int *tadOnlyShapeInfo = reinterpret_cast<int *>(tadShapeInfo);
 	int *tadOffset = reinterpret_cast<int *>(tadOffsets);
